src/third_party/blink/web_tests/accessibility/name-calc-presentational.html

<!DOCTYPE HTML>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/run-after-layout-and-paint.js"></script>

<div class="container">
  <div tabIndex=0 role="link" id="link1">
    I
    <img src="hidden.jpg" alt="do not" role="presentation" />
    like ice cream.
  </div>
</div>

<script>
test_after_layout_and_paint(function(t) {
    var axLink1 = accessibilityController.accessibleElementById("link1");
    assert_equals(axLink1.name, "I like ice cream.");
}, "Presentational element ignored in name calculation..");
</script>
